requirements.txt
flask
flask-cors
faster-whisper
python-dotenv
openai